+++ src/tests/fixtures/arguments/PositionalArgDeprecatedOptional.ts
@@ -1,0 +1,16 @@
+/** @gqlInput */
+type Greeting = {
+  name: string;
+  salutation: string;
+};
+
+/** @gqlType */
+export default class SomeType {
+  /** @gqlField */
+  hello(
+    /** @deprecated Unused! */
+    greeting?: Greeting | null,
+  ): string {
+    return `Hullo`;
+  }
+}