+++ src/tests/fixtures/arguments/PositionalArgOptional.ts
@@ -1,0 +1,13 @@
+/** @gqlInput */
+type Greeting = {
+  name: string;
+  salutation: string;
+};
+
+/** @gqlType */
+export default class SomeType {
+  /** @gqlField */
+  hello(greeting?: string | null): string {
+    return `${greeting ?? "Hello"} World`;
+  }
+}