+++ src/tests/fixtures/arguments/PositionalArgWithDefault.ts
@@ -1,0 +1,7 @@
+/** @gqlType */
+export default class SomeType {
+  /** @gqlField */
+  hello(greeting: string = "Hello"): string {
+    return `${greeting} World`;
+  }
+}