+++ src/tests/fixtures/input_types/ContextReferencedInInputType.invalid.ts
@@ -1,0 +1,9 @@
+/** @gqlContext */
+type SomeType = {
+  hello: string;
+};
+
+/** @gqlInput */
+type MyInputType = {
+  someField: SomeType;
+};