+++ src/tests/fixtures/resolver_context/ContextValueTypedAsString.invalid.ts
@@ -1,7 +1,0 @@
-/** @gqlType */
-export class SomeType {
-  /** @gqlField */
-  greeting(ctx: string): string {
-    return ctx;
-  }
-}