+++ src/transforms/resolveResolverParams.ts
@@ -1,14 +1,27 @@
-import { DefinitionNode, FieldDefinitionNode, Kind, visit } from "graphql";
-import { TypeContext } from "../TypeContext";
+import {
+  DefinitionNode,
+  FieldDefinitionNode,
+  InputValueDefinitionNode,
+  Kind,
+  visit,
+} from "graphql";
+import { TypeContext, UNRESOLVED_REFERENCE_NAME } from "../TypeContext";
 import { err, ok } from "../utils/Result";
 import {
   DiagnosticsResult,
   FixableDiagnosticWithLocation,
-  gqlErr,
+  gqlRelated,
+  tsErr,
 } from "../utils/DiagnosticError";
-import { loc } from "../utils/helpers";
+import { loc, nullThrows } from "../utils/helpers";
+import {
+  NamedFieldParam,
+  PositionalFieldParam,
+  Unresolved,
+  UnresolvedResolverParam,
+} from "../metadataDirectives.js";
 import * as E from "../Errors";
-import { UnresolvedResolverParam } from "../metadataDirectives.js";
+import { GraphQLConstructor } from "../GraphQLConstructor";
 
 export function resolveResolverParams(
   ctx: TypeContext,
@@ -20,9 +33,11 @@ export function resolveResolverParams(
 
 class ResolverParamsResolver {
   ctx: TypeContext;
+  gql: GraphQLConstructor;
   errors: FixableDiagnosticWithLocation[] = [];
   constructor(ctx: TypeContext) {
     this.ctx = ctx;
+    this.gql = new GraphQLConstructor();
   }
 
   resolve(
@@ -44,42 +59,93 @@ class ResolverParamsResolver {
     if (field.resolverParams == null) {
       return field;
     }
-    const nextResolverParams: UnresolvedResolverParam[] =
-      field.resolverParams.map((param) => {
-        return this.transformParam(param);
-      });
-    return { ...field, resolverParams: nextResolverParams };
+    // Resolve all the params individually
+    const resolverParams: UnresolvedResolverParam[] = field.resolverParams.map(
+      (param) => this.transformParam(param),
+    );
+
+    // Now we check to see if the params are a valid combination...
+    const args = resolverParams.find(
+      (param): param is NamedFieldParam =>
+        param.kind === "named" && param.name === "args",
+    );
+    const positionalArgs: PositionalFieldParam[] = resolverParams.filter(
+      (param) => param.kind === "positionalArg",
+    );
+
+    if (args != null && positionalArgs.length > 0) {
+      this.errors.push(
+        tsErr(nullThrows(args.sourceNode), E.positionalArgAndArgsObject(), [
+          gqlRelated(
+            loc(positionalArgs[0].inputDefinition),
+            "Positional GraphQL argument defined here",
+          ),
+        ]),
+      );
+      return field;
+    }
+
+    const fieldArgs: InputValueDefinitionNode[] =
+      field.arguments == null ? [] : [...field.arguments];
+
+    // Add any positional args to the field's arguments
+    for (const positionalArg of positionalArgs) {
+      fieldArgs.push(positionalArg.inputDefinition);
+    }
+
+    return { ...field, arguments: fieldArgs, resolverParams };
   }
 
   transformParam(param: UnresolvedResolverParam): UnresolvedResolverParam {
     switch (param.kind) {
       case "named":
+      case "positionalArg":
         return param;
       case "unresolved": {
-        const resolved = this.ctx.gqlNameDefinitionForGqlName(
-          param.namedTypeNode.name,
-        );
-        if (resolved.kind === "ERROR") {
-          this.errors.push(resolved.err);
-          return param;
-        }
-        switch (resolved.value.kind) {
-          case "CONTEXT":
-            return { kind: "named", name: "context" };
-          case "INFO":
-            return { kind: "named", name: "info" };
-          default: {
-            this.errors.push(
-              gqlErr(loc(param.namedTypeNode), E.invalidResolverParamType()),
-            );
+        const unwrappedType = this.gql.nullableType(param.inputDefinition.type);
+        if (
+          unwrappedType.kind === "NamedType" &&
+          unwrappedType.name.value === UNRESOLVED_REFERENCE_NAME
+        ) {
+          const resolved = this.ctx.gqlNameDefinitionForGqlName(
+            unwrappedType.name,
+          );
+          if (resolved.kind === "ERROR") {
+            this.errors.push(resolved.err);
             return param;
           }
+          switch (resolved.value.kind) {
+            case "CONTEXT":
+              return { kind: "named", name: "context" };
+            case "INFO":
+              return { kind: "named", name: "info" };
+            default: {
+              // We'll assume it's supposed to be a positional arg.
+              return this.resolveToPositionalArg(param) ?? param;
+            }
+          }
         }
+        // This can't be a context or info param so we'll assume it's supposed
+        // to be a positional arg.
+        return this.resolveToPositionalArg(param) ?? param;
       }
       default: {
         const _exhaustive: never = param;
         throw new Error("Unexpected param kind");
       }
     }
+  }
+  resolveToPositionalArg(unresolved: Unresolved): PositionalFieldParam | null {
+    if (unresolved.inputDefinition.name.kind === "ERROR") {
+      this.errors.push(unresolved.inputDefinition.name.err);
+      return null;
+    }
+    return {
+      kind: "positionalArg",
+      inputDefinition: {
+        ...unresolved.inputDefinition,
+        name: unresolved.inputDefinition.name.value,
+      },
+    };
   }
 }