+++ website/docs/gratsSample.grats.ts
@@ -21,7 +21,7 @@ class User {
   name: string = "Alice";
 
   /** @gqlField */
-  greeting(args: { salutation: string }): string {
-    return `${args.salutation}, ${this.name}`;
+  greeting(salutation: string): string {
+    return `${salutation}, ${this.name}`;
   }
 }